# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SocialForge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Tenant {
  id: number;
  name: string;
  plan: string;
  aiModel: string;
  createdAt: string;
}

export type TenantSettingsPlan = typeof TenantSettingsPlan[keyof typeof TenantSettingsPlan];


export const TenantSettingsPlan = {
  free: 'free',
  pro: 'pro',
  business: 'business',
} as const;

export interface TenantSettings {
  /** @minLength 1 */
  name?: string;
  plan?: TenantSettingsPlan;
  /** @minLength 1 */
  aiModel?: string;
}

export interface Usage {
  captions: number;
  images: number;
  periodStart: string;
}

export interface PlanLimits {
  /** Monthly caption generation limit. -1 means unlimited. */
  captions: number;
  /** Monthly image generation limit. -1 means unlimited. */
  images: number;
  /** Max brand kits. -1 means unlimited. */
  brandKits: number;
  /** Max scheduled posts. -1 means unlimited. */
  scheduledPosts: number;
}

export interface MeProfile {
  tenant: Tenant;
  usage: Usage;
  limits: PlanLimits;
}

export interface Plan {
  id: string;
  name: string;
  priceLabel: string;
  limits: PlanLimits;
  features: string[];
}

export interface BrandKit {
  id: number;
  name: string;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  voice: string;
  hashtags: string[];
  /** @nullable */
  logoPath?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BrandKitInput {
  /** @minLength 1 */
  name: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  voice?: string;
  hashtags?: string[];
  /** @nullable */
  logoPath?: string | null;
}

export interface BrandKitUpdate {
  /** @minLength 1 */
  name?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  voice?: string;
  hashtags?: string[];
  /** @nullable */
  logoPath?: string | null;
}

export interface ContentItem {
  id: number;
  title: string;
  caption: string;
  /** @nullable */
  imagePath?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  platform: string;
  status: string;
  /** @nullable */
  brandKitId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type ContentInputStatus = typeof ContentInputStatus[keyof typeof ContentInputStatus];


export const ContentInputStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
} as const;

export interface ContentInput {
  /** @minLength 1 */
  title: string;
  caption?: string;
  /** @nullable */
  imagePath?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  platform?: string;
  status?: ContentInputStatus;
  /** @nullable */
  brandKitId?: number | null;
}

export type ContentUpdateStatus = typeof ContentUpdateStatus[keyof typeof ContentUpdateStatus];


export const ContentUpdateStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
} as const;

export interface ContentUpdate {
  /** @minLength 1 */
  title?: string;
  caption?: string;
  /** @nullable */
  imagePath?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  platform?: string;
  status?: ContentUpdateStatus;
  /** @nullable */
  brandKitId?: number | null;
}

export interface CaptionRequest {
  /**
     * Topic or idea for the post.
     * @minLength 1
     */
  prompt: string;
  platform?: string;
  /** @nullable */
  brandKitId?: number | null;
  tone?: string;
}

export interface CaptionResult {
  caption: string;
  hashtags: string[];
}

export type ImageRequestSize = typeof ImageRequestSize[keyof typeof ImageRequestSize];


export const ImageRequestSize = {
  '1024x1024': '1024x1024',
  '1536x1024': '1536x1024',
  '1024x1536': '1024x1536',
} as const;

export interface ImageRequest {
  /** @minLength 1 */
  prompt: string;
  size?: ImageRequestSize;
  /** @nullable */
  brandKitId?: number | null;
}

export interface ImageResult {
  imagePath: string;
  b64Json: string;
}

export interface ScheduledPost {
  id: number;
  contentItemId: number;
  platform: string;
  scheduledAt: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface ScheduleInput {
  contentItemId: number;
  /** @minLength 1 */
  platform: string;
  scheduledAt: string;
}

export type ScheduleUpdateStatus = typeof ScheduleUpdateStatus[keyof typeof ScheduleUpdateStatus];


export const ScheduleUpdateStatus = {
  pending: 'pending',
  published: 'published',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface ScheduleUpdate {
  /** @minLength 1 */
  platform?: string;
  scheduledAt?: string;
  status?: ScheduleUpdateStatus;
}

export interface ConnectedAccount {
  id: number;
  platform: string;
  accountName: string;
  status: string;
  createdAt: string;
}

export type AccountInputPlatform = typeof AccountInputPlatform[keyof typeof AccountInputPlatform];


export const AccountInputPlatform = {
  instagram: 'instagram',
  facebook: 'facebook',
  linkedin: 'linkedin',
  youtube: 'youtube',
} as const;

export interface AccountInput {
  platform: AccountInputPlatform;
  /** @minLength 1 */
  accountName: string;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). */
  objectPath: string;
  metadata?: UploadUrlRequest;
}