# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brand-kits:
    get:
      operationId: listBrandKits
      tags: [brand-kits]
      summary: List brand kits
      responses:
        "200":
          description: List of brand kits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandKit"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createBrandKit
      tags: [brand-kits]
      summary: Create a brand kit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandKitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandKit"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /brand-kits/{id}:
    get:
      operationId: getBrandKit
      tags: [brand-kits]
      summary: Get a brand kit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand kit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandKit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateBrandKit
      tags: [brand-kits]
      summary: Update a brand kit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandKitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandKit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteBrandKit
      tags: [brand-kits]
      summary: Delete a brand kit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List content items
      responses:
        "200":
          description: List of content items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createContent
      tags: [content]
      summary: Create a content item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get a content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update a content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete a content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /ai/generate-caption:
    post:
      operationId: generateCaption
      tags: [ai]
      summary: Generate an AI caption with hashtags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptionRequest"
      responses:
        "200":
          description: Generated caption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptionResult"
        "402":
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /ai/generate-image:
    post:
      operationId: generateImage
      tags: [ai]
      summary: Generate an AI image and persist it to storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageRequest"
      responses:
        "200":
          description: Generated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResult"
        "402":
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List connected social accounts
      responses:
        "200":
          description: List of connected accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectedAccount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Connect a social account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedAccount"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /accounts/{id}:
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Disconnect a social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    BrandKit:
      type: object
      required:
        - id
        - name
        - primaryColor
        - secondaryColor
        - accentColor
        - voice
        - hashtags
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        voice:
          type: string
        hashtags:
          type: array
          items:
            type: string
        logoPath:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BrandKitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        voice:
          type: string
        hashtags:
          type: array
          items:
            type: string
        logoPath:
          type: ["string", "null"]
    BrandKitUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        voice:
          type: string
        hashtags:
          type: array
          items:
            type: string
        logoPath:
          type: ["string", "null"]
    ContentItem:
      type: object
      required:
        - id
        - title
        - caption
        - platform
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        caption:
          type: string
        imagePath:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        platform:
          type: string
        status:
          type: string
        brandKitId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContentInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        caption:
          type: string
        imagePath:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        platform:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
        brandKitId:
          type: ["integer", "null"]
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        caption:
          type: string
        imagePath:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        platform:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published]
        brandKitId:
          type: ["integer", "null"]
    CaptionRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          description: Topic or idea for the post.
        platform:
          type: string
        brandKitId:
          type: ["integer", "null"]
        tone:
          type: string
    CaptionResult:
      type: object
      required: [caption, hashtags]
      properties:
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
    ImageRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
        brandKitId:
          type: ["integer", "null"]
    ImageResult:
      type: object
      required: [imagePath, b64Json]
      properties:
        imagePath:
          type: string
        b64Json:
          type: string
...
    ConnectedAccount:
      type: object
      required: [id, platform, accountName, status, createdAt]
      properties:
        id:
          type: integer
        platform:
          type: string
        accountName:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    AccountInput:
      type: object
      required: [platform, accountName]
      properties:
        platform:
          type: string
          enum: [instagram, facebook, linkedin, youtube]
        accountName:
          type: string
          minLength: 1
...