# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SocialForge API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInputPlatform } from './accountInputPlatform';

export interface AccountInput {
  platform: AccountInputPlatform;
  /** @minLength 1 */
  accountName: string;
}