# lib/api-zod/src/generated/types/captionRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SocialForge API
 * OpenAPI spec version: 0.1.0
 */

export interface CaptionRequest {
  /**
     * Topic or idea for the post.
     * @minLength 1
     */
  prompt: string;
  platform?: string;
  /** @nullable */
  brandKitId?: number | null;
  tone?: string;
}